Update test-action.yml
Update test-action.yml

Update test-action.yml

feat: Add gpt4 option when using --file  (#48)

<!-- Prefix the PR title with 'feat:' to increment the minor version, or
'fix:' to increment the patch version -->

<!-- Describe what changes this PR introduces and their effects -->
<!-- Update documentation if there are any changes to existing features
or APIs -->

<!-- Describe how you tested your changes and how reviewers can test
them -->
<!-- Ensure all code can be tested offline (unmarked tests) and online
(marked as 'integration') -->

<!-- Add any other context about your changes here -->

<!-- ## Breaking Changes -->

<!-- List any breaking changes introduced by this PR and their
implications and include '!' in the prefix. -->

---------


Update Version to 0.5.0 [no ci]

Update pyproject.toml

Update python.yml

Update python.yml

Update pyproject.toml

pin version to >=3.8.1 based on langchain

Update python.yml

Update pyproject.toml

Update pyproject.toml

Update pyproject.toml

docs: Update README.md

Update README.md

Add fast mode option for GPT querying and update AzureChatOpenAI usage

Refactor: Load Azure context in _ask, add max_tokens to _get_engine, and count tokens in prompt

Refactor _ask.py: move Llama Index related code to _llama_index.py and add test

Refactor _llama_index.py for better readability and fix fast parameter usage

Add large model option and improve engine selection in GPT review

Update src/gpt_review/_ask.py


Refactor _llama_index.py: swap _ask_doc and AzureGPT35Turbo class positions

Add tests and pytest import for llama_indexer

Update test_llama_indexer to use fast mode in _ask_doc call

Add docstrings and return type annotations in _ask.py and _llama_index.py

Improve docstring for _count_tokens and refactor variable names in _document_indexer

Rename test_llama_indexer.py to test_llama_index.py

Add directory and repository support to _ask_doc and refactor GPT call functions

Remove duplicate 'large' parameter in _ask function

Remove AzureGPT35Turbo class and related functions from _llama_index.py

Update document loading order in _llama_index.py

Add fast and large model options to _request_goal function

Add fast and large options to _request_goal function.

Add options for GPT-4 and large prompts to git commit message generation.

"Add argument for using gpt-4 in git commit command"

Add required_exts parameter to _ask_doc function.
