+++ .github/workflows/test-action.yml
@@ -28,8 +28,8 @@ jobs:
           pip install -e .
           gpt github review \
             --access-token $GITHUB_TOKEN \
-            --pull-request $PATCH_PR \
-            --repository $PATCH_REPO
+            --pull-request $PR_NUMBER \
+            --repository $REPOSITORY_NAME
         env:
           GIT_COMMIT_HASH: ${{ github.event.pull_request.head.sha }}
           GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
@@ -40,3 +40,15 @@ jobs:
           REPOSITORY_NAME: ${{ github.repository }}
           AZURE_OPENAI_API: ${{ secrets.AZURE_OPENAI_API }}
           AZURE_OPENAI_API_KEY: ${{ secrets.AZURE_OPENAI_API_KEY }}
+          LINK: "https://github.com/${{ github.repository }}/pull/${{ github.event.pull_request.number }}"
+          FILE_SUMMARY: false
+          TEST_SUMMARY: false
+          BUG_SUMMARY: false
+          RISK_SUMMARY: false
+          RISK_ROLLBACK: false
+          RSIK_BREAKING: false
+          RISK_FLAGGED: false
+          SUMMARY_CONFIG: false
+          SUMMARY_SCHEMA: false
+          SUMMARY_FLAGS: false
+          SUMMARY_INCIDENTS: false
