+++ pyproject.toml
@@ -14,12 +14,12 @@ classifiers = [
     "Intended Audience :: Developers",
     "License :: OSI Approved :: MIT License",
     "Programming Language :: Python :: 3 :: Only",
-    "Programming Language :: Python :: 3.7",
     "Programming Language :: Python :: 3.8",
     "Programming Language :: Python :: 3.9",
-    "Programming Language :: Python :: 3.10"
+    "Programming Language :: Python :: 3.10",
+    "Programming Language :: Python :: 3.11"
 ]
-requires-python = ">=3.7"
+requires-python = ">=3.8.1"
 dynamic = ["version"]
 dependencies = [
   'azure-identity',
@@ -30,7 +30,9 @@ dependencies = [
   'knack',
   'openai',
   'requests',
-  "pyyaml",
+  'pyyaml',
+  'typing_extensions; python_version <= "3.10"',
+  'transformers; python_version <= "3.8"'
 ]
 
 [project.optional-dependencies]
