+++ src/gpt_review/__init__.py
@@ -5,4 +5,4 @@
 """Easy GPT CLI"""
 from __future__ import annotations
 
-__version__ = "0.4.0"
+__version__ = "0.5.0"
