+++ src/gpt_review/_git.py
@@ -4,6 +4,7 @@ import os
 from typing import Dict
 
 from knack import CLICommandsLoader
+from knack.arguments import ArgumentsContext
 from knack.commands import CommandGroup
 from git.repo import Repo
 
@@ -39,7 +40,7 @@ def _diff() -> str:
     return Repo.init(_find_git_dir()).git.diff(None, cached=True)
 
 
-def _commit_message() -> str:
+def _commit_message(fast: bool = False, large: bool = False) -> str:
     """
     Create a commit message with GPT.
 
@@ -53,17 +54,17 @@ Create a short, single-line, git commit message for these changes
     diff = _diff()
     logging.debug("Diff: %s", diff)
 
-    return _request_goal(diff, goal)
+    return _request_goal(diff, goal, fast=fast, large=large)
 
 
-def _commit() -> Dict[str, str]:
+def _commit(gpt4: bool = False, large: bool = False) -> Dict[str, str]:
     """
     Run git commit with a commit message generated by GPT.
 
     Returns:
         response (dict): The response from GPT-4.
     """
-    message = _commit_message()
+    message = _commit_message(fast=not gpt4, large=large)
     logging.debug("Commit Message: %s", message)
     repo = Repo.init(_find_git_dir())
     commit = repo.git.commit(message=message)
@@ -77,3 +78,19 @@ class GitCommandGroup(GPTCommandGroup):
     def load_command_table(loader: CLICommandsLoader) -> None:
         with CommandGroup(loader, "git", "gpt_review._git#{}") as group:
             group.command("commit", "_commit", is_preview=True)
+
+    @staticmethod
+    def load_arguments(loader: CLICommandsLoader) -> None:
+        with ArgumentsContext(loader, "git commit") as args:
+            args.argument(
+                "gpt4",
+                help="Use gpt-4 for prompts instead of gpt-35-turbo.",
+                default=False,
+                action="store_true",
+            )
+            args.argument(
+                "large",
+                help="Use gpt-4-32k for prompts.",
+                default=False,
+                action="store_true",
+            )
